fix(eval): measure the RCM parser that ships, not a private copy of it
_classify_rcm_event existed twice: inlined in radio_agent, which ships, and ported into strategy/eval/nlp.py, which is what the published coverage measured. The harness docstring asserted the two were identical.

Over the real 1515-row 2025 RCM corpus they disagreed on 427 messages, 28.2 percent. The shipped side is the correct one on SAFETY CAR IN THIS LAP, because #305's fix for the FIA end-of-neutralisation message never crossed into the port, so what was being measured was the older parser. There was also a structural divergence: RCMEvent has no sector field while the port branched on one.

The consequence reached a published number. documents/eval_reports/nlp.md carried rcm coverage 0.9868, which reproduces the port exactly, while the parser that actually ships scores 0.9644 on the same corpus. The report over-stated the shipped parser by 2.24 points, upward, on a row whose status column read reproduced.

The port existed for a real reason: importing radio_agent loads every NLP model. But the classifier is pure string rules with no model behind it, so the fix is to move it rather than copy it. src/f1_strat_manager/rcm_events.py now holds RCMEvent, the flag maps, the keyword families and classify_rcm_event. radio_agent imports them and the harness is a thin field-name adapter over the same function.

Verified: the two now agree on 1515 of 1515 rows, down from 427 disagreements; the regenerated report publishes 0.9644; 22 RCM/radio/NLP tests pass; ruff check and format clean repo-wide.

The sector question is deliberately left open rather than settled here. The persisted schema carries it, the shipped parser decides sector-scoping from scope alone, and which behaviour is right is a separate question from making the harness measure what ships.

Closes #632
